perf: added signal to abort pending connections
