+++ src/components/standalone/dashboard/IpsServiceCard.vue
@@ -5,18 +5,8 @@ import { getStandaloneRoutePrefix } from '@/lib/router'
 import { useI18n } from 'vue-i18n'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { faCheck, faXmark } from '@fortawesome/free-solid-svg-icons'
-import { useQuery } from '@tanstack/vue-query'
-import { ubusCall } from '@/lib/standalone/ubus'
 import { DASHBOARD_REFRESH_INTERVAL } from '@/composables/useDashboardOverview'
-
-type SnortStatusResponse = {
-  data: {
-    status: {
-      enabled: boolean
-      events: number
-    }
-  }
-}
+import { useIpsStatus } from '@/composables/useIpsStatus'
 
 const { t } = useI18n()
 
@@ -25,12 +15,7 @@ const {
   isPending,
   isError,
   error
-} = useQuery({
-  queryKey: ['dashboard', 'ips', 'status'],
-  queryFn: ({ signal }) => ubusCall<SnortStatusResponse>('ns.snort', 'status', {}, { signal }),
-  select: (res) => res.data.status,
-  refetchInterval: DASHBOARD_REFRESH_INTERVAL
-})
+} = useIpsStatus({ refetchInterval: DASHBOARD_REFRESH_INTERVAL })
 
 const errorTitle = computed(() => (isError.value ? t('standalone.ips.failed_to_fetch_info') : ''))
 const errorDescription = computed(() => (isError.value ? t(getAxiosErrorMessage(error.value)) : ''))
